feat: S24

+++ poly_forum/lib/screens/company/offers/get/components/delete_offer_btn.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:poly_forum/cubit/company/navigation/company_get_user_cubit.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:poly_forum/cubit/company/offer/company_get_offer_cubit.dart';
 import 'package:poly_forum/cubit/company/offer/company_offer_cubit.dart';
 import 'package:poly_forum/data/models/offer_model.dart';
+import 'package:poly_forum/screens/shared/components/modals/confirmation_modal.dart';
+import 'package:poly_forum/screens/shared/components/modals/modal_return_enum.dart';
 import 'package:poly_forum/screens/shared/components/row_btn.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:poly_forum/utils/constants.dart';
 import 'package:top_snackbar_flutter/custom_snack_bar.dart';
 import 'package:top_snackbar_flutter/top_snack_bar.dart';
@@ -32,22 +33,21 @@ class DeleteOfferBtn extends StatelessWidget {
               padding: kTopSnackBarPadding,
               child: const CustomSnackBar.error(
                 message:
-                    "Un problème est survenue, la sauvegarde pas été effectuée...",
+                    "Un problème est survenu, l'offre n'a pas pu être supprimée...",
               ),
             ),
           );
         } else if (state is CompanyOfferLoaded) {
+          BlocProvider.of<CompanyGetOfferCubit>(context).deleteLocalOffer(offer);
           showTopSnackBar(
             context,
             Padding(
               padding: kTopSnackBarPadding,
               child: const CustomSnackBar.success(
-                message: "La sauvegarde a été faite avec succès !",
+                message: "L'offre a été supprimée avec succès.",
               ),
             ),
           );
-          BlocProvider.of<CompanyGetOfferCubit>(context)
-              .deleteLocalOffer(offer);
         }
       },
       builder: (context, state) {
@@ -62,7 +62,19 @@ class DeleteOfferBtn extends StatelessWidget {
         return RowBtn(
           text: "Supprimer cette offre",
           onPressed: () {
-            BlocProvider.of<CompanyOfferCubit>(context).deleteOffer(offer);
+            showDialog(
+                context: context,
+                builder: (BuildContext context) {
+                  return ConfirmationModal(
+                      title: "Supprimer une offre",
+                      description: "Vous êtes sur le point de supprimer l'offre ${offer.name}, en êtes-vous sûr ?"
+                  );
+                }
+            ).then((value) {
+              if (value == ModalReturn.confirm) {
+                BlocProvider.of<CompanyOfferCubit>(context).deleteOffer(offer);
+              }
+            });
           },
           color: kRedButton,
         );
